module_four/requirements.txt
alembic==1.17.0
annotated-types==0.7.0
anyio==4.11.0
asttokens==3.0.0
bcrypt==5.0.0
beautifulsoup4==4.14.0
cachetools==6.2.2
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
cloudpickle==3.1.2
colorama==0.4.6
comm==0.2.3
contourpy==1.3.3
cryptography==46.0.3
cycler==0.12.1
debugpy==1.8.17
decorator==5.2.1
distro==1.9.0
dotenv==0.9.9
et_xmlfile==2.0.0
executing==2.2.1
fastapi==0.119.1
filelock==3.19.1
fonttools==4.59.2
gdown==5.2.0
git-filter-repo==2.47.0
google-auth==2.43.0
google-genai==1.52.0
greenlet==3.2.4
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
imbalanced-learn==0.14.0
ipykernel==6.30.1
ipython==9.5.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jiter==0.12.0
joblib==1.5.2
jupyter_client==8.6.3
jupyter_core==5.8.1
kagglehub==0.3.13
kiwisolver==1.4.9
lightgbm==4.6.0
llvmlite==0.45.1
Mako==1.3.10
MarkupSafe==3.0.3
matplotlib==3.10.7
matplotlib-inline==0.2.1
nest-asyncio==1.6.0
numba==0.62.1
numpy==2.3.3
ollama==0.6.1
openai==2.8.1
openpyxl==3.1.5
packaging==25.0
pandas==2.3.2
parso==0.8.5
pillow==11.3.0
platformdirs==4.4.0
prompt_toolkit==3.0.52
psutil==7.1.0
pure_eval==0.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.12.3
pydantic_core==2.41.4
Pygments==2.19.2
PyJWT==2.10.1
PyMySQL==1.1.2
pyparsing==3.2.4
PySocks==1.7.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
pywin32==311
PyYAML==6.0.3
pyzmq==27.1.0
requests==2.32.5
rsa==4.9.1
scikit-learn==1.7.2
scipy==1.16.2
seaborn==0.13.2
shap==0.49.1
six==1.17.0
slicer==0.0.8
sniffio==1.3.1
soupsieve==2.8
SQLAlchemy==2.0.44
stack-data==0.6.3
starlette==0.48.0
tenacity==9.1.2
threadpoolctl==3.6.0
tornado==6.5.2
tqdm==4.67.1
traitlets==5.14.3
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
uvicorn==0.38.0
wcwidth==0.2.13
websockets==15.0.1
xgboost==3.1.1
